+++ src/app/(header)/layout.tsx
@@ -1,21 +1,14 @@
 import { Header } from '@/components';
-import { AuthProvider } from '@/features/auth/components';
-import { cookies } from 'next/headers';
 
-export default async function HeaderLayout({
+export default async function AuthHeaderLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const cookieStore = await cookies();
-  const refreshToken = cookieStore.get('refreshToken')?.value;
-
   return (
-    <AuthProvider refreshToken={refreshToken}>
-      <Header />
-      <div className="relative px-6 min-h-[calc(100vh-4rem)] h-full">
-        {children}
-      </div>
-    </AuthProvider>
+    <>
+      <Header title="로그인 및 회원가입" />
+      <div className="px-6 min-h-[calc(100vh-4rem)] h-full">{children}</div>
+    </>
   );
 }
