+++ src/app/(menu-header)/projects/[id]/not-found.tsx
