--- src/features/auth/components/auth-header/index.ts
@@ -1,1 +1,0 @@
-export * from './AuthHeader';
