+++ src/features/auth/components/index.ts
@@ -1,4 +1,3 @@
-export * from './auth-header';
 export * from './auth-provider';
 export * from './login-button';
 export * from './login-callback-content';
